Ed_Agent/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-dotenv
gunicorn

# LangChain + AI
langchain
langchain-community
langchain-openai
openai
chromadb

# PDF / PPTX
pdfminer.six>=20221105
python-pptx>=0.6.21
